refactor: move webapi model discovery
